auth/requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
click==8.1.7
fastapi==0.115.5
greenlet==3.1.1
h11==0.14.0
idna==3.10
pydantic==2.10.2
pydantic_core==2.27.1
pyodbc==5.2.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
typing_extensions==4.12.2
uvicorn==0.32.1
python-dotenv==1.0.1 
passlib==1.7.4 
python-jose[cryptography]==3.3.0
passlib==1.7.4
python-multipart==0.0.6